+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Changes to Joyride
 ## [Unreleased]
 
 - [Add `joyride.core/*file*`](https://github.com/BetterThanTomorrow/joyride/issues/5)
-- Change name of command *Run Script* -> *Run Clojure Code*
+- [Change name of command *Run Script* -> *Run Clojure Code*]((https://github.com/BetterThanTomorrow/joyride/issues/20))
+- [Introduce User scripts](https://github.com/BetterThanTomorrow/joyride/issues/5)
 
 ## [0.0.2] - 2022-04-27