+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Changes to Joyride
 
 ## [Unreleased]
 
+## [0.0.4] - 2022-05-05
+
+- [API for starting the nREPL server](https://github.com/BetterThanTomorrow/joyride/issues/28)
+- [`when` clause contexts for nREPL server](https://github.com/BetterThanTomorrow/joyride/issues/29)
+
+## [0.0.3] - 2022-05-02
+
 - [Add `joyride.core/*file*`](https://github.com/BetterThanTomorrow/joyride/issues/5)
 - [Change name of command *Run Script* -> *Run Clojure Code*]((https://github.com/BetterThanTomorrow/joyride/issues/20))
 - [Introduce User scripts](https://github.com/BetterThanTomorrow/joyride/issues/5)